+++ evalml/pipelines/components/estimators/classifiers/lightgbm_classifier.py
@@ -79,7 +79,8 @@ class LightGBMClassifier(Estimator):
         X = infer_feature_types(X)
         cat_cols = list(X.select('category').columns)
         X = _convert_woodwork_types_wrapper(X.to_dataframe())
-
+        if fit:
+            self.input_feature_names = list(X.columns)
         X_encoded = _rename_column_names_to_numeric(X)
         rename_cols_dict = dict(zip(X.columns, X_encoded.columns))
         cat_cols = [rename_cols_dict[col] for col in cat_cols]
@@ -105,9 +106,11 @@ class LightGBMClassifier(Estimator):
         return y_encoded
 
     def fit(self, X, y=None):
+        X_encoded = _convert_to_woodwork_structure(X)
         X_encoded = self._encode_categories(X, fit=True)
         y_encoded = self._encode_labels(y)
-        return super().fit(X_encoded, y_encoded)
+        self._component_obj.fit(X_encoded, y_encoded)
+        return self
 
     def predict(self, X):
         X_encoded = self._encode_categories(X)