+++ evalml/pipelines/components/estimators/estimator.py
@@ -9,7 +9,7 @@ from evalml.utils import _convert_woodwork_types_wrapper, infer_feature_types
 class Estimator(ComponentBase):
     """A component that fits and predicts given data.
 
-    To implement a new Transformer, define your own class which is a subclass of Transformer, including
+    To implement a new Estimator, define your own class which is a subclass of Estimator, including
     a name and a list of acceptable ranges for any parameters to be tuned during the automl search (hyperparameters).
     Define an `__init__` method which sets up any necessary state and objects. Make sure your `__init__` only
     uses standard keyword arguments and calls `super().__init__()` with a parameters dict. You may also override the
@@ -26,6 +26,25 @@ class Estimator(ComponentBase):
     @abstractmethod
     def supported_problem_types(cls):
         """Problem types this estimator supports"""
+
+    def __init__(self, parameters=None, component_obj=None, random_state=0, **kwargs):
+        self.input_feature_names = None
+        super().__init__(parameters=parameters, component_obj=component_obj, random_state=random_state, **kwargs)
+
+    def _manage_woodwork(self, X, y=None):
+        """Function to convert the input and target data to Pandas data structures."""
+        X = _convert_to_woodwork_structure(X)
+        X = _convert_woodwork_types_wrapper(X.to_dataframe())
+        if y is not None:
+            y = _convert_to_woodwork_structure(y)
+            y = _convert_woodwork_types_wrapper(y.to_series())
+        return X, y
+
+    def fit(self, X, y=None):
+        X, y = self._manage_woodwork(X, y)
+        self.input_feature_names = list(X.columns)
+        self._component_obj.fit(X, y)
+        return self
 
     def predict(self, X):
         """Make predictions using selected features.