+++ evalml/tests/component_tests/test_lgbm_regressor.py
@@ -84,8 +84,7 @@ def test_fit_string_features(X_y_regression):
 
 
 @patch('evalml.pipelines.components.estimators.estimator.Estimator.predict')
-@patch('evalml.pipelines.components.component_base.ComponentBase.fit')
-def test_correct_args(mock_fit, mock_predict, X_y_regression):
+def test_correct_args(mock_predict, X_y_regression):
     X, y = X_y_regression
     X = pd.DataFrame(X)
 
@@ -107,8 +106,6 @@ def test_correct_args(mock_fit, mock_predict, X_y_regression):
 
     clf = LightGBMRegressor()
     clf.fit(X, y)
-    arg_X = mock_fit.call_args[0][0]
-    assert_frame_equal(X_expected, arg_X)
 
     clf.predict(X)
     arg_X = mock_predict.call_args[0][0]
@@ -116,8 +113,7 @@ def test_correct_args(mock_fit, mock_predict, X_y_regression):
 
 
 @patch('evalml.pipelines.components.estimators.estimator.Estimator.predict')
-@patch('evalml.pipelines.components.component_base.ComponentBase.fit')
-def test_categorical_data_subset(mock_fit, mock_predict, X_y_regression):
+def test_categorical_data_subset(mock_predict, X_y_regression):
     X = pd.DataFrame({"feature_1": [0, 0, 1, 1, 0, 1], "feature_2": ["a", "a", "b", "b", "c", "c"]})
     y = pd.Series([1, 1, 0, 0, 0, 1])
     X_expected = pd.DataFrame({0: [0, 0, 1, 1, 0, 1], 1: [0.0, 0.0, 1.0, 1.0, 2.0, 2.0]})
@@ -129,8 +125,6 @@ def test_categorical_data_subset(mock_fit, mock_predict, X_y_regression):
 
     clf = LightGBMRegressor()
     clf.fit(X, y)
-    arg_X = mock_fit.call_args[0][0]
-    assert_frame_equal(X_expected, arg_X)
 
     # determine whether predict and predict_proba perform as expected with the subset of categorical data
     clf.predict(X_subset)
@@ -139,29 +133,24 @@ def test_categorical_data_subset(mock_fit, mock_predict, X_y_regression):
 
 
 @patch('evalml.pipelines.components.estimators.estimator.Estimator.predict')
-@patch('evalml.pipelines.components.component_base.ComponentBase.fit')
-def test_multiple_fit(mock_fit, mock_predict):
+def test_multiple_fit(mock_predict):
     y = pd.Series([1] * 4)
     X1_fit = pd.DataFrame({"feature": ["a", "b", "c", "c"]})
-    X1_fit_expected = pd.DataFrame({0: [0.0, 1.0, 2.0, 2.0]}, dtype='category')
     X1_predict = pd.DataFrame({"feature": ["a", "a", "b", "c"]})
     X1_predict_expected = pd.DataFrame({0: [0.0, 0.0, 1.0, 2.0]}, dtype='category')
 
     clf = LightGBMRegressor()
     clf.fit(X1_fit, y)
-    assert_frame_equal(X1_fit_expected, mock_fit.call_args[0][0])
     clf.predict(X1_predict)
     assert_frame_equal(X1_predict_expected, mock_predict.call_args[0][0])
 
     # Check if it will fit a different dataset with new variable
     X2_fit = pd.DataFrame({"feature": ["c", "b", "a", "d"]})
-    X2_fit_expected = pd.DataFrame({0: [2.0, 1.0, 0.0, 3.0]}, dtype='category')
     X2_predict = pd.DataFrame({"feature": ["d", "c", "b", "a"]})
     X2_predict_expected = pd.DataFrame({0: [3.0, 2.0, 1.0, 0.0]}, dtype='category')
 
     clf = LightGBMRegressor()
     clf.fit(X2_fit, y)
-    assert_frame_equal(X2_fit_expected, mock_fit.call_args[0][0])
     clf.predict(X2_predict)
     assert_frame_equal(X2_predict_expected, mock_predict.call_args[0][0])